6558 improve import logging (#6586)
* #6558 Improved logging for import when the temp folder is not available or accessible

* #6558 Small improvements for data import

* #6558 Fixed conflicts

* #6558 Added new tempDirectory check for CSV imports

* #6558 Fixed unit tests

+++ sormas-ui/src/main/java/de/symeda/sormas/ui/importer/ImportReceiver.java
@@ -6,6 +6,7 @@ import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.OutputStream;
+import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Date;
 import java.util.function.Consumer;
@@ -73,9 +74,13 @@ public class ImportReceiver implements Receiver, StartedListener, SucceededListe
 
 		final FileOutputStream fos;
 		try {
+			Path tempDirectory = Paths.get(FacadeProvider.getConfigFacade().getTempFilesPath());
+			if (!tempDirectory.toFile().exists() || !tempDirectory.toFile().canWrite()) {
+				throw new FileNotFoundException("Temp directory doesn't exist or cannot be accessed");
+			}
 			String newFileName = ImportExportUtils.TEMP_FILE_PREFIX + fileNameAddition + DateHelper.formatDateForExport(new Date()) + "_"
 				+ DataHelper.getShortUuid(UserProvider.getCurrent().getUuid()) + ".csv";
-			file = new File(Paths.get(FacadeProvider.getConfigFacade().getTempFilesPath()).resolve(newFileName).toString());
+			file = new File(tempDirectory.resolve(newFileName).toString());
 			fos = new FileOutputStream(file);
 		} catch (FileNotFoundException e) {
 			file = null;
